move mapQueryRowsToValueSets to utils
