+++ query-connector/src/app/database-service.ts
@@ -1,5 +1,5 @@
 "use server";
-import { Pool, PoolConfig, QueryResultRow } from "pg";
+import { Pool, PoolConfig,  } from "pg";
 import {
   Bundle,
   OperationOutcome,
@@ -140,55 +140,6 @@ export const executeDefaultQueryCreation = async () => {
     console.error("Could not cross-index default queries", error);
     throw error;
   }
-};
-
-/**
- * Maps the results returned from the DIBBs value set and coding system database
- * into a collection of value sets, each containing one or more Concepts build out
- * of the coding information in the DB.
- * @param rows The Rows returned from the DB Query.
- * @returns A list of ValueSets, which hold the Concepts pulled from the DB.
- */
-export const mapQueryRowsToValueSets = async (rows: QueryResultRow[]) => {
-  // Create groupings of rows (each of which is a single Concept) by their ValueSet ID
-  const vsIdGroupedRows = rows.reduce((conceptsByVSId, r) => {
-    if (!(r["valueset_id"] in conceptsByVSId)) {
-      conceptsByVSId[r["valueset_id"]] = [];
-    }
-    conceptsByVSId[r["valueset_id"]].push(r);
-    return conceptsByVSId;
-  }, {});
-
-  // Each "prop" of the struct is now a ValueSet ID
-  // Iterate over them to create formal Concept Groups attached to a formal VS
-  const valueSets = Object.keys(vsIdGroupedRows).map((vsID) => {
-    const conceptGroup: QueryResultRow[] = vsIdGroupedRows[vsID];
-    const valueSet: ValueSet = {
-      valueSetId: conceptGroup[0]["valueset_id"],
-      valueSetVersion: conceptGroup[0]["version"],
-      valueSetName: conceptGroup[0]["valueset_name"],
-      // External ID might not be defined for user-defined valuesets
-      valueSetExternalId: conceptGroup[0]["valueset_external_id"]
-        ? conceptGroup[0]["valueset_external_id"]
-        : undefined,
-      author: conceptGroup[0]["author"],
-      system: conceptGroup[0]["code_system"],
-      ersdConceptType: conceptGroup[0]["type"]
-        ? conceptGroup[0]["type"]
-        : undefined,
-      dibbsConceptType: conceptGroup[0]["dibbs_concept_type"],
-      includeValueSet: conceptGroup.find((c) => c["include"]) ? true : false,
-      concepts: conceptGroup.map((c) => {
-        return {
-          code: c["code"],
-          display: c["display"],
-          include: c["include"],
-        };
-      }),
-    };
-    return valueSet;
-  });
-  return valueSets;
 };
 
 /*
