+++ query-connector/src/app/query/components/selectQuery/queryHooks.ts
@@ -6,8 +6,8 @@ import {
 } from "@/app/constants";
 import {
   getSavedQueryByName,
-  mapQueryRowsToValueSets,
 } from "@/app/database-service";
+import   { mapQueryRowsToValueSets } from "@/app/utils"
 import { UseCaseQuery, UseCaseQueryResponse } from "@/app/query-service";
 import { Patient } from "fhir/r4";
