+++ query-connector/src/app/queryBuilding/components/ConditionSelection.tsx
@@ -20,7 +20,7 @@ import { BuildStep } from "@/app/constants";
 import { FormError } from "../buildFromTemplates/page";
 
 import { ValueSet } from "@/app/constants";
-import { mapQueryRowsToValueSets } from "@/app/database-service";
+import { mapQueryRowsToValueSets } from "@/app/utils";
 
 type ConditionSelectionProps = {
   fetchedConditions: CategoryNameToConditionOptionMap;
@@ -98,7 +98,7 @@ export const ConditionSelection: React.FC<ConditionSelectionProps> = ({
 
       if (isSubscribed) {
         setFetchedConditions(
-          mapFetchedDataToFrontendStructure(categoryToConditionArrayMap),
+          mapFetchedDataToFrontendStructure(categoryToConditionArrayMap)
         );
       }
     }
@@ -110,6 +110,7 @@ export const ConditionSelection: React.FC<ConditionSelectionProps> = ({
   }, []);
 
   async function getValueSetsForSelectedConditions() {
+    // return array of promises
     const conditionIds = Object.entries(selectedConditions)
       .map(([_, conditionObj]) => {
         return Object.keys(conditionObj);
@@ -117,6 +118,8 @@ export const ConditionSelection: React.FC<ConditionSelectionProps> = ({
       .flatMap((ids) => ids);
 
     const ConditionValueSets: ConditionIdToValueSetArray = {};
+    // const joinInsertsPromiseArray = generateValuesetConceptJoinSqlPromises(vs);
+    // const joinInsertResults = await Promise.allSettled(joinInsertsPromiseArray);
 
     for (const id of conditionIds) {
       const results: ValueSet[] =
@@ -136,7 +139,7 @@ export const ConditionSelection: React.FC<ConditionSelectionProps> = ({
   }
 
   async function handleCreateQueryClick(
-    event: React.MouseEvent<HTMLButtonElement>,
+    event: React.MouseEvent<HTMLButtonElement>
   ) {
     event.preventDefault();
     setLoading(true);
@@ -155,7 +158,7 @@ export const ConditionSelection: React.FC<ConditionSelectionProps> = ({
     <div
       className={classNames(
         "bg-gray-5 margin-top-4 ",
-        styles.conditionTemplateContainer,
+        styles.conditionTemplateContainer
       )}
     >
       <div className="display-flex flex-justify flex-align-end margin-bottom-3 width-full">
@@ -179,7 +182,7 @@ export const ConditionSelection: React.FC<ConditionSelectionProps> = ({
           id="conditionTemplateSearch"
           placeholder="Search conditions"
           className={classNames(
-            "maxw-mobile margin-x-auto margin-top-0 margin-bottom-4",
+            "maxw-mobile margin-x-auto margin-top-0 margin-bottom-4"
           )}
           onChange={(e) => {
             e.preventDefault();
